Infra: fix broken unidriver import
